Further refactoring towards capabilities
