+++ psqlpack/src/model/extension.rs
@@ -1,6 +1,6 @@
 use connection::Connection;
 use errors::{PsqlpackResult, PsqlpackErrorKind};
-use model::{Capabilities, DefinableCatalog, Package};
+use model::{Capabilities, DefinableCatalog, MetaInfo, Package, SourceInfo};
 use semver::Semver;
 
 use slog::Logger;
@@ -20,22 +20,23 @@ impl Extension {
 
         trace!(log, "Connecting to database");
         let db_conn = connection.connect_database()?;
-
+        let meta = MetaInfo::new(SourceInfo::Extension);
         let context = capabilities.with_context(self);
-        let schemas = context.schemata(&db_conn, connection.database());
-        /*
+        let schemas = context.query_schemata(&db_conn, connection.database())?;
+        let types = context.query_types(&db_conn)?;
+        dbtry!(db_conn.finish());
+
         let mut package = Package {
             meta,
             extensions: Vec::new(),
-            functions,
-            indexes,
+            functions: Vec::new(), // TODO
+            indexes: Vec::new(), // TODO
             schemas,
             scripts: Vec::new(),
-            tables,
+            tables: Vec::new(), // TODO
             types,
         };
         package.promote_primary_keys_to_table_constraints();
-        Ok(package)*/
-        panic!("")
+        Ok(package)
     }
 }
